update as per requirements

+++ database/migrations/2024_10_18_083014_create_t_stock_transfer_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('t_stock_transfer', function (Blueprint $table) {
             $table->id();
-            $table->integer('transfer_id');
+            $table->unsignedBigInteger('transfer_id');
             $table->string('godown_from');
             $table->string('godown_to');
             $table->date('transfer_date');
