+++ database/migrations/2024_10_18_083057_create_t_assembly_products_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('t_assembly_products', function (Blueprint $table) {
             $table->id();
-            $table->integer('assembly_id');
+            $table->unsignedBigInteger('assembly_id');
             $table->integer('product_id');
             $table->string('product_name');
             $table->integer('quantity');
